working on stretch case, factored description to own module
