+++ public/client/similarview.jsx
@@ -1,0 +1,3 @@
+//write a customers who viewed this item also viewed
+  //~ 7 other items
+  //stretch case: arrows that allow scrolling to other items in the database
